week13

+++ week-13/doctor_appointment/appointments/views.py
@@ -3,23 +3,35 @@ from appointments.models import *
 from appointments.serializer import *
 from rest_framework.decorators import APIView
 from rest_framework.response import Response
-from rest_framework.parsers import JSONParser
 from datetime import datetime
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
+from rest_framework.authentication import TokenAuthentication
+from appointments.permissions import AppointmentListPermission, AppointmentDetailsPermission
+
+class MyTokenAuthentication(TokenAuthentication):
+    keyword = "Bearer"
 # Create your views here.
 class DoctorList(APIView):
+    authentication_classes = [MyTokenAuthentication]
+    permission_classes = [IsAuthenticated]
     def get(self, request):
         doctors = Doctor.objects.all()
         serializer = DoctorSerializer(doctors, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 class PatientList(APIView):
+    authentication_classes = [MyTokenAuthentication]
+    permission_classes = [IsAuthenticated]
     def get(self, request):
         patients = Patient.objects.all()
         serializer = PatientSerializer(patients, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 class AppointmentList(APIView):
+    authentication_classes = [MyTokenAuthentication]
+    permission_classes = [AppointmentListPermission]
+
     def get(self, request):
         appointments = Appointment.objects.all()
         serializers = AppointmentSerializer(appointments, many=True)
@@ -41,12 +53,14 @@ class AppointmentList(APIView):
         appointment = CreatAppointmentSerializer(data=data)
 
         if appointment.is_valid():
-            appointment.save()
+            appointment.save(created_by=request.user)
             return Response(appointment.data, status=status.HTTP_201_CREATED)
         else:
             return Response(appointment.errors, status=status.HTTP_200_OK)
     
 class AppointmentDetails(APIView):
+    authentication_classes = [MyTokenAuthentication]
+    permission_classes = [AppointmentDetailsPermission]
     
     def get(self, request, app_id):
         try:
@@ -57,21 +71,27 @@ class AppointmentDetails(APIView):
             return Response(data={'error' : 'Appointment not found'}, status=status.HTTP_404_NOT_FOUND)
     
     def patch(self, request, app_id):
+        
         try:
             appo = Appointment.objects.get(id=app_id)
-            serializers = CreatAppointmentSerializer(appo, data=request.data)
-            if serializers.is_valid():
-                serializers.save()
-                return Response(serializers.data, status=status.HTTP_200_OK)
-            else:
-                return Response(serializers.errors, status=status.HTTP_400_BAD_REQUEST)
-        except:
+        except Appointment.DoesNotExist:
             return Response(data={'error' : 'Appointment not found'}, status=status.HTTP_404_NOT_FOUND)
+        
+        self.check_object_permissions(request, appo)
+        serializers = CreatAppointmentSerializer(appo, data=request.data)
+        if serializers.is_valid():
+            serializers.save()
+            return Response(serializers.data, status=status.HTTP_200_OK)
+        else:
+            return Response(serializers.errors, status=status.HTTP_400_BAD_REQUEST)
         
     def delete(self, request, app_id):
         try:
             appo = Appointment.objects.get(id=app_id)
-            appo.delete()
-            return Response(data={'message' : "Appointment Deleted"}, status=status.HTTP_200_OK)
-        except:
+        except Appointment.DoesNotExist:
             return Response(data={'error' : 'Appointment not found'}, status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(request, appo)
+        appo.delete()
+        return Response(data={'message' : "Appointment Deleted"}, status=status.HTTP_200_OK)
+
+
